requirements.txt
absl-py==0.7.1
aequitas==0.38.0
alabaster==0.7.12
alibi==0.2.2
altair==3.2.0
anchor-exp==0.0.0.5
argh==0.26.2
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
base58==1.0.3
basescript==0.2.9
beautifulsoup4==4.7.1
bleach==3.1.0
blinker==1.4
blis==0.2.4
blist==1.3.6
boto3==1.9.253
botocore==1.12.253
bs4==0.0.1
catboost==0.18
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cloudpickle==1.0.0
colorama==0.3.9
cycler==0.10.0
cymem==2.0.2
dash==0.39.0
dash-core-components==0.44.0
dash-cytoscape==0.1.1
dash-html-components==0.14.0
dash-renderer==0.20.0
dash-table==3.6.0
dash-table-experiments==0.6.0
dask==1.2.2
decorator==4.4.0
deeputil==0.2.7
defusedxml==0.6.0
Dijkstar==2.4.0
dill==0.2.9
diskarray==0.1.9
docutils==0.14
dominate==2.3.5
ds-lime==0.1.1.27
edx-sphinx-theme==1.4.0
entrypoints==0.3
enum-compat==0.0.3
enum34==1.1.6
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.4.0
future==0.17.1
gast==0.2.2
gevent==1.4.0
graphviz==0.10.1
greenlet==0.4.15
grpcio==1.20.1
h5py==2.9.0
html5lib==1.0.1
httplib2==0.12.3
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.17
interpret==0.1.7
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lightgbm==2.2.3
lime==0.1.1.33
Markdown==3.1
markdown2==2.3.5
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mock==3.0.5
more-itertools==7.0.0
multiprocess==0.70.7
murmurhash==1.0.2
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
numpy==1.16.4
ohio==0.4.0
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pathlib2==2.3.3
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
pidigits==0.5
Pillow==6.0.0
pkg-resources==0.0.0
pkginfo==1.5.0.1
plac==0.9.6
plotly==3.10.0
pluggy==0.12.0
preshed==2.0.1
progressbar2==3.47.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.6.2
ptyprocess==0.6.0
PuLP==1.6.10
py==1.8.0
pycebox==0.0.1
pyCeterisParibus==0.5.1
pydotplus==2.0.2
pygam==0.8.0
Pygments==2.4.0
pyparsing==2.4.0
PyPDF2==1.26.0
pyrsistent==0.15.2
pysbrl==0.4.1
PyScaffold==3.1
pytest==4.6.2
pytest-runner==5.1
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2018.3
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.1
qii-tool==0.1.3
readme-renderer==24.0
recordclass==0.5
reportlab==3.5.23
repoze.lru==0.7
requests==2.21.0
requests-toolbelt==0.9.1
retrying==1.3.3
s3transfer==0.2.1
SALib==1.3.6
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.1
seaborn==0.9.0
shap==0.28.5
six==1.11.0
skater==1.1.2
sklearn==0.0
skope-rules==1.0.0
snowballstemmer==1.2.1
soupsieve==1.9.1
spacy==2.1.4
Sphinx==2.1.0
sphinx-bootstrap-theme==0.6.5
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.4
srsly==0.0.5
streamlit==0.48.1
structlog==18.1.0
tabulate==0.8.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
testpath==0.4.2
themis-ml==0.0.4
thinc==7.0.4
toml==0.10.0
toolz==0.9.0
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
tsp==0.0.6
twine==1.13.0
tzlocal==2.0.0
urllib3==1.24.3
validators==0.14.0
visitor==0.1.3
wasabi==0.2.2
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
wordcloud==1.3.1
xgboost==0.90
xhtml2pdf==0.2.2
zipp==0.5.1
